+++ clingon/clingon.py
@@ -23,7 +23,7 @@ import os
 import sys
 import textwrap
 
-__version__ = '0.1.4a3'
+__version__ = '0.1.4rc1'
 DEBUG = False
 DELAY_EXECUTION = False
 SYSTEM_EXIT_ERROR_CODE = 1